# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "paths": {
        "/api/v1/feed/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "add feed",
                "parameters": [
                    {
                        "description": "FeedAddForm",
                        "name": "feedAddForm",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.FeedAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/all/{userid}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "get feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "userid",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/delete/{id}": {
            "get": {
                "tags": [
                    "feed"
                ],
                "summary": "delete feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/get/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "get feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/pagination": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "get feed",
                "parameters": [
                    {
                        "type": "string",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/public/pagination": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "get feed",
                "parameters": [
                    {
                        "type": "string",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/refresh/{id}": {
            "get": {
                "tags": [
                    "feed"
                ],
                "summary": "add feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/feed/update": {
            "get": {
                "tags": [
                    "feed"
                ],
                "summary": "add feed",
                "parameters": [
                    {
                        "description": "数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/form.FeedUpdateForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/feed/rss/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "get feed",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "feed"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据"
                },
                "message": {
                    "description": "消息",
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "description": "状态",
                    "type": "boolean"
                }
            }
        },
        "form.FeedAddForm": {
            "type": "object",
            "properties": {
                "public": {
                    "type": "boolean"
                },
                "toLang": {
                    "type": "string"
                },
                "translateDescription": {
                    "type": "boolean"
                },
                "translateTitle": {
                    "type": "boolean"
                },
                "translatorID": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "form.FeedUpdateForm": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "fromLang": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "toLang": {
                    "type": "string"
                },
                "translateDescription": {
                    "type": "boolean"
                },
                "translateTitle": {
                    "type": "boolean"
                },
                "translatorId": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "Bearer Token 格式：`Bearer {token}`\nJWT Authorization header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}